feat: add generic support for displaying Shell
