+++ samples/Playground/Playground.Shared/Views/AppRoot.xaml.cs
@@ -1,12 +1,14 @@
-﻿using Uno.Toolkit.UI;
+using Uno.Toolkit.UI;
 
 namespace Playground.Views;
 
-public sealed partial class AppRoot : UserControl
+public sealed partial class AppRoot : UserControl, IContentControlProvider
 {
 	public ExtendedSplashScreen SplashScreen => Splash;
 	public AppRoot()
 	{
 		this.InitializeComponent();
 	}
+
+	public ContentControl ContentControl => Splash;
 }
