+++ src/Uno.Extensions.Hosting.UI/ApplicationBuilder.cs
@@ -4,6 +4,8 @@ internal record ApplicationBuilder(Application App, LaunchActivatedEventArgs Arg
 {
 	private readonly List<Action<IHostBuilder>> _delegates = new List<Action<IHostBuilder>>();
 
+	public IDictionary<object, object> Properties { get; } = new Dictionary<object, object>();
+
 	public Window Window { get; } =
 #if NET6_0_OR_GREATER && WINDOWS && !HAS_UNO
 		new Window();
