+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.UI/App.xaml.cs
@@ -30,7 +30,9 @@ public sealed partial class App : Application
 			.Localization()
 			.Serialization()
 			.Services()
-			.Navigation();
+			.Navigation()
+			// Add navigation support for toolkit controls such as TabBar and NavigationView
+			.UseToolkitNavigation();;
 		_window = builder.Window;
 
 		_host = await builder.ShowAsync<Shell>();
